compiles, "runs"
couldn't get SWO to work for... unknow reasons.  Blobbed in RTT.  prints
like it's running, but duh, just discovered this is not for the f429
nucleo, it's for some f429igt6 with the HS port connected. So... it
doesn't really do anything like what I hoped for.

What board? not sure, it's not the waveshare CORE429I either, that's
only go the FS port connected.
Possibly the
https://stm32-base.org/boards/STM32F429IGT6-LC-Technology.html that
implies it has the HS port connected? but not sure whwere you could get
it?

Still, keep all the work done.

new files are from regen from cubemx